requirements.txt
admin==0.0.1
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
autopep8==1.5.7
beautifulsoup4==4.9.3
billiard==3.6.3.0
boto3==1.15.18
botocore==1.18.18
bower==0.0.0
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
charts==0.4.6
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
cloudinary==1.25.0
cryptography==3.3.1
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.1.4
django-admin-charts==0.24.1
django-admin-filter==0.6
django-admin-interface==0.16.1
django-admin-tools==0.9.1
django-advanced-filters==1.3.0
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-bower==5.2.0
django-braces==1.14.0
django-celery==3.3.1
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cloudinary-storage==0.3.0
django-colorfield==0.4.1
django-cors-headers==3.6.0
django-crontab==0.7.1
django-fcm==0.1.1
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-heroku==0.3.1
django-import-export==2.0.2
django-jsonfield==1.4.1
django-liststyle==0.1
django-money==1.3.1
django-nested-admin==3.3.2
django-nvd3==0.9.7
django-phone-field==1.8.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.1.0
django-rest-registration==0.6.0
django-restframework==0.0.1
django-simple-history==2.12.0
django-storages==1.8
django-timezone-field==4.1.2
django-user-activity-log2==0.0.24
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djcacheutils==3.0.0
et-xmlfile==1.0.1
fcm-django==0.3.7
gunicorn==20.0.4
idna==2.10
jdcal==1.4.1
Jinja2==2.10.3
jmespath==0.10.0
joblib==0.17.0
kombu==4.6.11
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==1.1.1
numpy==1.19.3
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.3
pandas==1.1.3
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
py-moneyed==0.8.0
pycodestyle==2.7.0
pycparser==2.20
pyfcm==1.4.9
PyJWT==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-memcached==1.59
python-monkey-business==1.0.0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
razorpay==1.2.0
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.3.0
s3transfer==0.2.1
scikit-learn==0.24.1
simplejson==3.17.2
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
tablib==1.1.0
threadpoolctl==2.1.0
toml==0.10.2
Unidecode==1.2.0
Unipath==1.1
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.3.0
